% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shades.R
\name{palette_show}
\alias{palette_show}
\title{Visualise the output of a palette generating function}
\usage{
palette_show(f)
}
\arguments{
\item{f}{The palette generating function}
}
\description{
Visualise the output of a palette generating function
}
