% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shades.R
\name{shades}
\alias{shades}
\title{Palettes varying only in alpha}
\usage{
shades(name)
}
\arguments{
\item{name}{a colour name}
}
\value{
a palette generating function
}
\description{
Palettes varying only in alpha
}
