% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shades.R
\name{shades}
\alias{shades}
\title{Palettes varying only in alpha}
\usage{
shades(col)
}
\arguments{
\item{col}{a colour specification recognised by col2rgb}
}
\value{
a palette generating function
}
\description{
Palettes varying only in alpha
}
